Implement function tests auto-generation and linking in database
Added the ability to automatically generate tests for functions stored in the database. Also, linked the generated tests to their corresponding functions within the database. Scripts for generating tests and writing them to a file can be found in 'agent/core.py'. Modifications were also made in 'code_management' modules to enable linking of tests to functions and adding the newly generated tests to the database.