Add database reset function and improve test linking
Added a function to reset the database and made improvements to the test linking, introducing support for identifying whether a test relates to a class method. These enhancements improve test management by correctly identifying and linking tests to their corresponding functions or class methods. For example, a new field has been added to record if the test is for a class method, and this additional metadata is used during linking process.

Signed-off-by: sandbox <sandbox@simibrum.com>